refactor: fix typo in doctor-entity.mjs file name
